+++ backend/src/main/java/com/celuveat/video/query/dao/VideoQueryDaoSupport.java
@@ -1,11 +1,11 @@
-package com.celuveat.video.query;
+package com.celuveat.video.query.dao;
 
 import com.celuveat.restaurant.command.domain.Restaurant;
 import com.celuveat.video.command.domain.Video;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VideoQueryRepository extends JpaRepository<Video, Long>, VideoEntityManagerQueryRepository {
+public interface VideoQueryDaoSupport extends JpaRepository<Video, Long> {
 
     List<Video> findAllByRestaurantIdIn(List<Long> restaurantIds);
